# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация для  сервиса, хранящего пользователя и сегменты, в которых он состоит.",
        "title": "Dynamic service",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "paths": {
        "/add_user_to_segment": {
            "post": {
                "description": "Добавление сегментов пользователю.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Пользовательские сегменты",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserSegments"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/create_segment": {
            "post": {
                "description": "Создание сегмента.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Название сегмента",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Segment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Segment"
                        }
                    }
                }
            }
        },
        "/create_user": {
            "post": {
                "description": "Создание пользователя.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Пользователь",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/delete_segment": {
            "post": {
                "description": "Удаление сегмента.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Название сегмента",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Segment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/delete_user_from_segment": {
            "post": {
                "description": "Удаление сегментов у пользователя.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Пользовательские сегменты",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserSegments"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/show_segments": {
            "get": {
                "description": "Вывод списка активных сегментов у пользователя.",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Пользователь",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UserId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.UserId": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "controllers.UserSegments": {
            "type": "object",
            "properties": {
                "segments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "domain.Segment": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.User": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}